+++ Sources/NIO/BaseSocketChannel.swift
@@ -1310,3 +1310,13 @@ extension BaseSocketChannel {
         return SynchronousOptions(_channel: self)
     }
 }
+
+/// Execute the given function and synchronously complete the given `EventLoopPromise` (if not `nil`).
+func executeAndComplete<Value>(_ promise: EventLoopPromise<Value>?, _ body: () throws -> Value) {
+    do {
+        let result = try body()
+        promise?.succeed(result)
+    } catch let e {
+        promise?.fail(e)
+    }
+}
