+++ Tests/NIOTests/EventLoopTest.swift
@@ -12,6 +12,7 @@
 //
 //===----------------------------------------------------------------------===//
 import XCTest
+@testable import NIOCore
 @testable import NIO
 import Dispatch
 import NIOConcurrencyHelpers
